backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1
pandas>=2.2.3
numpy>=2.0.0
faker==33.3.0
